fix(ios): explicit CGFloat conversions, drop unverified pickups metric
Relying on SE-0307's implicit CGFloat/Double bridging in view builders is the
kind of thing that reads fine and then fails to type-check under a result
builder, so the four Double-to-CGFloat sites are now explicit.

Also drops UsageSummary.pickups: it read
segment.totalPickupsWithoutApplicationActivity, which I could not verify against
the DeviceActivityData API, and nothing rendered it.

+++ studytimer/Sources/App/Design/HoldToEndButton.swift
@@ -27,7 +27,7 @@ struct HoldToEndButton: View {
                 GeometryReader { geo in
                     Capsule()
                         .fill(Theme.red.opacity(0.28))
-                        .frame(width: geo.size.width * fill)
+                        .frame(width: geo.size.width * CGFloat(fill))
                 }
             }
